updated all uses of filesys, almost ready to go!

+++ src/utils/crc32.h
@@ -5,15 +5,19 @@
 // From https://gist.github.com/xobs/91a84d29152161e973d717b9be84c4d0
 // (not using fast version because we want small binary size)
 // Adapted such that original CRC can be passed in to continue calculation
-// Note this does NOT invert after calculation, this must be done by the user!
-inline static unsigned int crc32_continue(const uint8_t *message, uint16_t len,
+// WARNING: Note this does NOT invert after calculation, this must be done by
+// the user! To use, initialize crc to 0xFFFFFFFF, then call crc32_continue as
+// needed, then invert the final result (use the ~ operator).
+// This is equivalent to zlib's implementation, and should always be the same as
+// running in Python: import zlib zlib.crc32(data) Where data is a byte-string
+// (use bytes()) with the same data as message.
+inline static unsigned int crc32_continue(const uint8_t *message, size_t len,
                                           unsigned int crc)
 {
     size_t i;
-    unsigned int byte, /*crc,*/ mask;
+    unsigned int byte, mask;
 
     i = 0;
-    // crc = 0xFFFFFFFF;
     while (i < len)
     {
         byte = message[i]; // Get next byte.
@@ -29,7 +33,7 @@ inline static unsigned int crc32_continue(const uint8_t *message, uint16_t len,
     return crc; // Don't invert here - must be done by the user!
 }
 
-inline static unsigned int crc32(const uint8_t *message, uint16_t len)
+inline static unsigned int crc32(const uint8_t *message, size_t len)
 {
     return ~crc32_continue(message, len, 0xFFFFFFFF);
 }
